move CommandLineAnalyst to tetris package
